+++ native/java/org/jpype/javadoc/JavadocTransformer.java
@@ -97,7 +97,8 @@ public class JavadocTransformer
     {
       data.hr = true;
       parent.removeChild(node);
-    } else if (name.equals("pre"))
+    } else if (name.equals("pre") || // Javadoc pre-17
+        (name.equals("div") && e.getAttribute("class").equals("type-signature"))) // Javadoc 17+
     {
       DomUtilities.removeWhitespace(node);
       doc.renameNode(node, null, "signature");
@@ -124,10 +125,11 @@ public class JavadocTransformer
     String name = e.getTagName();
     Document doc = e.getOwnerDocument();
 
-    if (name.equals("h4"))
+    if (name.equals("h4") || name.equals("h3")) // h4 for Javadoc pre-17, h3 for Javadoc 17+
     {
       doc.renameNode(node, null, "title");
-    } else if (name.equals("pre"))
+    } else if (name.equals("pre") || // Javadoc pre-17
+          (name.equals("div") && (e.getAttribute("class").equals("member-signature")))) // Javadoc 17+
     {
       doc.renameNode(node, null, "signature");
       DomUtilities.traverseDFS(node, this::pass1, Node.ELEMENT_NODE, ws);