+++ native/java/org/jpype/pkg/JPypePackageManager.java
@@ -20,6 +20,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.file.DirectoryStream;
+import java.net.URLDecoder;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystemNotFoundException;
 import java.nio.file.FileSystems;
@@ -462,13 +463,34 @@ public class JPypePackageManager
     }
   }
 
-  // Java 8 windows bug https://bugs.java.com/bugdatabase/view_bug.do?bug_id=8131067
+  @SuppressWarnings("deprecation")
   private static URI toURI(Path path)
   {
     URI uri = path.toUri();
-    if (uri.getScheme().equals("jar") && uri.toString().contains("%2520"))
-      uri = URI.create("jar:" + uri.getRawSchemeSpecificPart().replaceAll("%25", "%"));
-    return uri;
+
+    try {
+      // Java 8 bug https://bugs.java.com/bugdatabase/view_bug.do?bug_id=8131067
+      // Zip file system provider returns doubly % encoded URIs. We resolve this
+      // by re-encoding the URI after decoding it.
+      uri = new URI(
+        uri.getScheme(),
+        URLDecoder.decode(uri.getSchemeSpecificPart()),
+        uri.getFragment()
+      );
+
+    // `toASCIIString` ensures the URI is URL encoded with only ascii
+    // characters. This avoids issues in `sun.nio.fs.UnixUriUtils.fromUri` that
+    // naively uses `uri.getRawPath()` despite the possibility that it contains
+    // non-ascii characters that will cause errors. By using `toASCIIString` and
+    // re-wrapping it in a URI object we ensure that the URI is properly
+    // encoded. See: https://github.com/jpype-project/jpype/issues/1194
+      return new URI(uri.toASCIIString());
+    } catch (Exception e) {
+      // This exception *should* never occur as we are re-encoding a valid URI.
+      // Throwing a runtime exception avoids java exception handling boilerplate
+      // for a situation that *should* never occur.
+      throw new RuntimeException("Failed to encode URI: " + uri, e);
+    }
   }
 //</editor-fold>
 }