+++ doc/CHANGELOG.rst
@@ -6,7 +6,11 @@ This changelog *only* contains changes from the *first* pypi release (0.5.4.3) o
 Latest Changes:
 
 - **1.5.1_dev0 - 2023-12-15**
+
+  - Added support for typing ``JArray`` (Java type only), e.g. ``JArray[java.lang.Object]`` ``"JArray[java.lang.Object]"``
+
   - Fixed uncaught exception while setting traceback causing issues in Python 3.11/3.12.
+
   - Use PEP-518 and PEP-660 configuration for the package, allowing editable and
     configurable builds using modern Python packaging tooling.
     Where before ``python setup.py --enable-tracing develop``, now can be done with
@@ -22,6 +26,12 @@ Latest Changes:
   - Update for tests for numpy 2.0.
 
   - Support of np.float16 conversion with arrays.
+
+  - Fixed a problem that caused ``dir(jpype.JPackage("mypackage"))`` to fail if
+    the class path contained non-ascii characters. See issue #1194.
+
+  - Fixed ``BufferOverflowException`` in ``JUnpickler`` when decoding
+    multiple Java objects.
 
 
 - **1.5.0 - 2023-04-03**